Fix issues 2 and 3
In this commmit, I transfer all the config options to the config.yaml
file, closing off issue #2.

I also attempt to fix issue #3 by keeping the Snakefile within the
package folder, and setting the working directory in the call to
snakemake.

In this implementation, the config.yaml file is put in the working
folder, but the Snakefile is never copied out of the package directory.
This makes a cleaner Snakefile, and means we are able to run both
locally and from within a container. The user should never have to
change anything in the Snakefile, but we should expose options
through the config.

Bump version to 0.1.1

Change extension for config file

Add toml dependency

Move variables from Snakemake.smk to config

Clean up

Put final output files in a list

This means that files are consistently passed on to the config template,
and then joined in the Snakefile if needed.

Change prefix for config file to j2

Add workdir to Snakemake command

That allows the working dir to be distinct from the dir where
the Snakefile is kept.

It also allows the user to call abriTAMR from a different
folder from where things will be run.

Effectively, this passes the workir parameter to Snakemake for
consistency with the rest of the script.

Add workdir suffix to output files

This allows for outputs files to be searched for explicitly in the
workdir rather than assume that the user is always running abriTAMR
from the workdir.

Fix missing reference to self

Add toml dependency

Add abritamr.log and Singularity image files

Add xlsxwriter dependency

Add xlsxwriter dependency

Initial commit

Bump version to 0.1.2

Make sure we have the right dependencies in place for MDU workflow